Refactor and add documentation
#20
