--- app/src/main/java/com/leondeklerk/starling/edit/BrushStyle.kt
@@ -1,10 +1,0 @@
-package com.leondeklerk.starling.edit
-
-data class BrushStyle(
-    val type: BrushType,
-    val hue: Float,
-    val saturation: Float,
-    val value: Float,
-    val size: Float,
-    val alpha: Float
-)
