--- app/src/main/java/com/leondeklerk/starling/edit/BrushType.kt
@@ -1,7 +1,0 @@
-package com.leondeklerk.starling.edit
-
-enum class BrushType {
-    PENCIL,
-    MARKER,
-    ERASER
-}
