--- app/src/main/java/com/leondeklerk/starling/edit/GradientType.kt
@@ -1,7 +1,0 @@
-package com.leondeklerk.starling.edit
-
-enum class GradientType {
-    HUE,
-    SATURATION,
-    VALUE
-}
