--- app/src/main/java/com/leondeklerk/starling/edit/Side.kt
@@ -1,9 +1,0 @@
-package com.leondeklerk.starling.edit
-
-enum class Side {
-    LEFT,
-    RIGHT,
-    TOP,
-    BOTTOM,
-    NONE
-}
