--- app/src/main/java/com/leondeklerk/starling/edit/TextStyle.kt
@@ -1,8 +1,0 @@
-package com.leondeklerk.starling.edit
-
-data class TextStyle(
-    val hue: Float,
-    val saturation: Float,
-    val value: Float,
-    val size: Float
-)
