+++ app/src/main/java/com/leondeklerk/starling/edit/crop/HandlerType.kt
@@ -1,5 +1,11 @@
-package com.leondeklerk.starling.edit
+package com.leondeklerk.starling.edit.crop
 
+/**
+ * Specifies the different types of touch handlers available.
+ * Each handler represents a specific point on a box.
+ * The corners are specified by left/right_top/bottom.
+ * The box type relates to a location within the box.
+ */
 enum class HandlerType {
     TOP,
     RIGHT,
