+++ app/src/main/java/com/leondeklerk/starling/edit/crop/Side.kt
@@ -1,0 +1,12 @@
+package com.leondeklerk.starling.edit.crop
+
+/**
+ * Enum class representing the different sides a object can be moving to.
+ */
+enum class Side {
+    LEFT,
+    RIGHT,
+    TOP,
+    BOTTOM,
+    NONE
+}
