+++ app/src/main/java/com/leondeklerk/starling/edit/draw/BrushType.kt
@@ -1,0 +1,10 @@
+package com.leondeklerk.starling.edit.draw
+
+/**
+ * Enum class representing the different types of available brushes
+ */
+enum class BrushType {
+    PENCIL,
+    MARKER,
+    ERASER
+}
