+++ app/src/main/java/com/leondeklerk/starling/edit/draw/GradientType.kt
@@ -1,0 +1,11 @@
+package com.leondeklerk.starling.edit.draw
+
+/**
+ * The different types of gradients.
+ * Used to differentiate between value space size for HSV color attributes.
+ */
+enum class GradientType {
+    HUE,
+    SATURATION,
+    VALUE
+}
