+++ app/src/main/java/com/leondeklerk/starling/edit/draw/TextStyleModal.kt
@@ -1,4 +1,4 @@
-package com.leondeklerk.starling.edit
+package com.leondeklerk.starling.edit.draw
 
 import android.content.DialogInterface
 import android.graphics.Color
@@ -12,7 +12,14 @@ import com.google.android.material.bottomsheet.BottomSheetBehavior
 import com.google.android.material.bottomsheet.BottomSheetDialog
 import com.google.android.material.bottomsheet.BottomSheetDialogFragment
 import com.leondeklerk.starling.databinding.ModalTextStyleBinding
+import com.leondeklerk.starling.views.ColorDotView
+import com.leondeklerk.starling.views.GradientSlider
 
+/**
+ * Modal responsible for selecting the different properties of a text style.
+ * Contains sliders for the HSV color attributes and size.
+ * Additionally includes a [ColorDotView] preview
+ */
 class TextStyleModal : BottomSheetDialogFragment() {
 
     private lateinit var binding: ModalTextStyleBinding
@@ -36,23 +43,44 @@ class TextStyleModal : BottomSheetDialogFragment() {
         super.onViewCreated(view, savedInstanceState)
         dialog?.setOnShowListener { dialog ->
             (dialog as BottomSheetDialog).let {
+                // Make sure the dialog cannot be dragged
                 val bottomSheet = it.findViewById<View>(com.google.android.material.R.id.design_bottom_sheet) as FrameLayout
                 val bottomSheetBehavior = BottomSheetBehavior.from(bottomSheet)
                 bottomSheetBehavior.isDraggable = false
             }
         }
 
-        setupTextViews()
+        setup()
     }
 
-    private fun setupTextViews() {
+    override fun onCancel(dialog: DialogInterface) {
+        super.onCancel(dialog)
+        onCloseListener?.invoke(getStyle())
+    }
+
+    /**
+     * Create a [TextStyle] from the current values
+     * @return a [TextStyle] based on the current values
+     */
+    fun getStyle(): TextStyle {
+        return TextStyle(hue, saturation, value, size)
+    }
+
+    /**
+     * Set up all listeners and the gradients for all the sliders.
+     * Responsible for creating the trackGradients of the [GradientSlider]s
+     */
+    private fun setup() {
+        // Create all the starting gradients and preview
         binding.sliderHue.trackGradient = getGradient(GradientType.HUE, 360f, 1f)
         binding.sliderSaturation.trackGradient = getGradient(GradientType.SATURATION, hue, 1f)
         binding.sliderValue.trackGradient = getGradient(GradientType.VALUE, hue, saturation)
         setPreview()
 
+        // Set up the slider listeners
         binding.sliderHue.addOnChangeListener { _, sliderValue, _ ->
             hue = sliderValue
+            // On a change of hue, update the other slider gradients
             binding.sliderSaturation.trackGradient = getGradient(GradientType.SATURATION, hue, 1f)
             binding.sliderValue.trackGradient = getGradient(GradientType.VALUE, hue, saturation)
             setPreview()
@@ -60,6 +88,7 @@ class TextStyleModal : BottomSheetDialogFragment() {
 
         binding.sliderSaturation.addOnChangeListener { _, sliderValue, _ ->
             saturation = sliderValue
+            // Create the new value gradient based on the saturation value
             binding.sliderValue.trackGradient = getGradient(GradientType.VALUE, hue, saturation)
             setPreview()
         }
@@ -75,28 +104,32 @@ class TextStyleModal : BottomSheetDialogFragment() {
         }
     }
 
-    override fun onCancel(dialog: DialogInterface) {
-        super.onCancel(dialog)
-        onCloseListener?.invoke(getStyle())
-    }
-
+    /**
+     * Based on the current color and size, apply the values to the [ColorDotView] preview.
+     */
     private fun setPreview() {
         val color = Color.HSVToColor(floatArrayOf(hue, saturation, value))
         binding.sizeColorPreview.setTextColor(color)
         binding.sizeColorPreview.textSize = size
     }
 
-    fun getStyle(): TextStyle {
-        return TextStyle(hue, saturation, value, size)
-    }
-
+    /**
+     * Create HSV gradients based on the type of HSV attribute and the current values.
+     * Based on the type, a range of different values is created (360 for hue, 100 for saturation/value)
+     * @param type: the type of HSV attribute this gradient represents
+     * @param hue the current hue value used in the saturation/value sliders
+     * @param saturation the current saturation used in the value slider
+     * @return a [GradientDrawable] representing the range of possible values for this type
+     */
     private fun getGradient(type: GradientType, hue: Float, saturation: Float): GradientDrawable {
+        // Determine the correct number of steps available
         val spaceSize = when (type) {
             GradientType.HUE -> 360
             GradientType.VALUE, GradientType.SATURATION -> 100
         }
 
         val colors = IntArray(spaceSize)
+        // Create the specific color for each step of the attribute space.
         for (i in 0 until spaceSize) {
             colors[i] = when (type) {
                 GradientType.HUE -> Color.HSVToColor(floatArrayOf(i.toFloat(), 1f, 1f))
@@ -105,10 +138,11 @@ class TextStyleModal : BottomSheetDialogFragment() {
             }
         }
 
+        // Create the actual gradient drawable based on the color list
         return GradientDrawable(GradientDrawable.Orientation.LEFT_RIGHT, colors)
     }
 
     companion object {
-        const val TAG = "TextStyleDialog"
+        const val TAG = "TextStyleModal"
     }
 }
