+++ app/src/main/java/com/leondeklerk/starling/extensions/Rect.kt
@@ -4,25 +4,22 @@ import android.graphics.PointF
 import android.graphics.Rect
 import androidx.core.graphics.contains
 import androidx.core.graphics.toPoint
-import java.security.InvalidParameterException
 
-fun Rect.shrinkBy(pixels: Int): Rect {
-    if (pixels <= width() / 2 && pixels <= height() / 2) {
-        return Rect(left + pixels, top + pixels, right - pixels, bottom - pixels)
-    } else {
-        throw InvalidParameterException()
-    }
-}
-
-fun Rect.enlargeBy(pixels: Int): Rect {
-    return Rect(left - pixels, top - pixels, right + pixels, bottom + pixels)
-}
-
+/**
+ * Enlarge a rectangle by a number of pixels in each direction.
+ * @param pixels: the number of pixels to enlarge by
+ * @return a new rect with increased size on each side
+ */
 fun Rect.enlargeBy(pixels: Float): Rect {
     val px = pixels.toInt()
     return Rect(left - px, top - px, right + px, bottom + px)
 }
 
+/**
+ * Check if a rectangle contains a point.
+ * @param point: the point to check for.
+ * @return if the point is within the rectangle or not.
+ */
 fun Rect.contains(point: PointF): Boolean {
     return (contains(point.toPoint()))
 }
