+++ app/src/main/res/layout/activity_image.xml
@@ -28,7 +28,7 @@
             android:layout_width="match_parent"
             android:layout_height="match_parent" />
 
-        <com.leondeklerk.starling.edit.InteractiveImageView
+        <com.leondeklerk.starling.views.InteractiveImageView
             android:visibility="visible"
             android:id="@+id/imageView"
             tools:srcCompat="@tools:sample/avatars"
