+++ app/src/main/res/layout/modal_text_style.xml
@@ -29,7 +29,7 @@
         app:layout_constraintEnd_toEndOf="@+id/text_title"
         android:layout_marginTop="8dp" />
 
-    <com.leondeklerk.starling.edit.GradientSlider
+    <com.leondeklerk.starling.views.GradientSlider
         android:id="@+id/slider_hue"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
@@ -44,7 +44,7 @@
         app:layout_constraintHorizontal_bias="0.5"
         app:layout_constraintBottom_toTopOf="@+id/slider_saturation" />
 
-    <com.leondeklerk.starling.edit.GradientSlider
+    <com.leondeklerk.starling.views.GradientSlider
         android:id="@+id/slider_saturation"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
@@ -59,7 +59,7 @@
         tools:trackGradient="@drawable/tools_gradient_saturation"
         android:layout_marginEnd="16dp" />
 
-    <com.leondeklerk.starling.edit.GradientSlider
+    <com.leondeklerk.starling.views.GradientSlider
         android:id="@+id/slider_value"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
