+++ app/src/main/res/layout/view_crop.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent">
 
-    <com.leondeklerk.starling.edit.CropOverlayView
+    <com.leondeklerk.starling.edit.crop.CropOverlayView
         android:id="@+id/overlay"
         android:layout_width="match_parent"
         android:layout_height="0dp"
