+++ app/src/main/res/layout/view_draw.xml
@@ -71,7 +71,7 @@
         app:layout_constraintStart_toEndOf="@+id/button_redo"
         app:layout_constraintHorizontal_bias="0.5" />
 
-    <com.leondeklerk.starling.edit.PaintView
+    <com.leondeklerk.starling.edit.draw.PaintView
         android:layout_width="0dp"
         android:layout_height="0dp"
         android:id="@+id/canvas"
