+++ app/src/main/res/layout/view_edit.xml
@@ -6,7 +6,7 @@
     android:background="@color/black"
     android:layout_height="match_parent">
 
-    <com.leondeklerk.starling.edit.InteractiveImageView
+    <com.leondeklerk.starling.views.InteractiveImageView
         android:layout_width="0dp"
         android:layout_height="0dp"
         tools:srcCompat="@tools:sample/avatars"
@@ -20,7 +20,7 @@
         android:layout_marginBottom="72dp"
         app:layout_constraintBottom_toTopOf="@+id/mode_divider" />
 
-    <com.leondeklerk.starling.edit.CropView
+    <com.leondeklerk.starling.edit.crop.CropView
         android:id="@+id/cropper"
         android:layout_width="match_parent"
         android:layout_height="0dp"
@@ -30,7 +30,7 @@
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintBottom_toTopOf="@id/mode_divider" />
 
-    <com.leondeklerk.starling.edit.DrawView
+    <com.leondeklerk.starling.edit.draw.DrawView
         android:visibility="invisible"
         tools:visibility="invisible"
         android:id="@+id/draw_overlay"
